Move admin PaginationController to ui
