+++ apps/admin/src/assets/index.ts
@@ -4,7 +4,6 @@ export { default as GoogleIcon } from './svg/icon/GoogleIcon';
 export { default as InputValueResetBtnIcon } from './svg/icon/InputValueResetBtnIcon';
 export { default as ModalCloseIcon } from './svg/icon/ModalCloseIcon';
 export { default as NotificationIcon } from './svg/icon/NotificationIcon';
-export { default as PaginationIcon } from './svg/icon/PaginationIcon';
 export { default as PlusIcon } from './svg/icon/PlusIcon';
 export { default as SeeMoreIcon } from './svg/icon/SeeMoreIcon';
 export { default as ToBackIcon } from './svg/icon/ToBackIcon';
