+++ packages/ui/src/assets/index.ts
@@ -1,1 +1,2 @@
 export { default as FileIcon } from './icons/FileIcon';
+export { default as PaginationIcon } from './icons/PaginationIcon';
