+++ packages/ui/src/components/PaginationController/index.stories.tsx
@@ -3,7 +3,7 @@ import PaginationController from '.';
 import type { Meta, StoryObj } from '@storybook/react';
 
 export default {
-  title: 'admin/PaginationController',
+  title: 'ui/PaginationController',
   component: PaginationController,
   parameters: {},
 } as Meta<typeof PaginationController>;
@@ -13,6 +13,6 @@ type Story = StoryObj<typeof PaginationController>;
 export const Primary: Story = {
   args: {
     totalPages: 5,
-    pageNumber: 0,
+    pageNumber: 1,
   },
 };
