+++ packages/ui/src/components/index.ts
@@ -3,3 +3,4 @@ export { default as DateComponent } from './DateComponent';
 export { default as FileButton } from './FileButton';
 export { default as LoadingCircle } from './LoadingCircle';
 export { default as WriterAndDate } from './WriterAndDate';
+export { default as PaginationController } from './PaginationController';
