+++ packages/ui/src/emotion.d.ts
@@ -1,0 +1,10 @@
+import '@emotion/react';
+
+import type { theme } from 'common';
+
+type ExtendedTheme = typeof theme;
+
+declare module '@emotion/react' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  interface Theme extends ExtendedTheme {}
+}
